labs/lab3/9771_Sachu/requirement.txt
 Flask
 Flask-PyMongo
 Flask-CORS
 certifi
